requirements.txt
asgiref==3.11.0
Django==4.2.27
djangorestframework==3.16.1
PyMySQL==1.1.2
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
